Changed type of IDs to Long on DTOs

+++ src/main/java/io/tackle/pathfinder/dto/AssessmentDto.java
@@ -45,14 +45,14 @@ public class AssessmentDto {
      */
     @JsonProperty("id")
     @JsonPropertyDescription("Assessment id")
-    private Integer id;
+    private Long id;
     /**
      * Application id
      * 
      */
     @JsonProperty("applicationId")
     @JsonPropertyDescription("Application id")
-    private Integer applicationId;
+    private Long applicationId;
     /**
      * Current status of the assessment
      *
